reorganized code into components, views, hooks, etc
