--- .env
@@ -1,1 +1,0 @@
-GOOGLE_MAPS_API_KEY="AIzaSyByVwAy15t8c4VoLJ-ZK1qBlPxh-I0_NUQ"
