+++ index.html
@@ -4,7 +4,7 @@
     <meta
       name="viewport"
       content="width=device-width, initial-scale=1.0, user-scalable=no" />
-    <title>Heatmap</title>
+    <title>US West Coast Earthquake Heatmap</title>
 
     <style>
       body {
