--- src/db.ts
@@ -1,11 +1,0 @@
-import { Pool } from 'pg';
-
-const pool = new Pool({
-  host: 'localhost',
-  port: 5432,
-  database: 'postgres',
-  user: 'tillmandickson',
-  password: '',
-});
-
-export default pool;
