+++ vite.config.js
@@ -1,11 +1,13 @@
 import {defineConfig, loadEnv} from 'vite';
 
 export default defineConfig(({mode}) => {
-  const {GOOGLE_MAPS_API_KEY = ''} = loadEnv(mode, process.cwd(), '');
+  const {VITE_GOOGLE_MAPS_API_KEY = ''} = loadEnv(mode, process.cwd(), '');
+  const {VITE_GOOGLE_MAPS_MAP_ID = ''} = loadEnv(mode, process.cwd(), '');
 
   return {
     define: {
-      'process.env.GOOGLE_MAPS_API_KEY': JSON.stringify(GOOGLE_MAPS_API_KEY)
+      'process.env.VITE_GOOGLE_MAPS_API_KEY': JSON.stringify(VITE_GOOGLE_MAPS_API_KEY),
+      'process.env.VITE_GOOGLE_MAPS_MAP_ID': JSON.stringify(VITE_GOOGLE_MAPS_MAP_ID)
     },
     resolve: {
       alias: {
